+++ Backend/Spring/Java/toby/src/main/java/com/poppo/toby/services/UserServiceTx.java
@@ -1,0 +1,34 @@
+package com.poppo.toby.services;
+
+import com.poppo.toby.domain.User;
+import org.springframework.transaction.PlatformTransactionManager;
+import org.springframework.transaction.TransactionStatus;
+import org.springframework.transaction.support.DefaultTransactionDefinition;
+
+public class UserServiceTx implements UserService {
+    private UserService userService;
+    private PlatformTransactionManager transactionManager;
+
+    public UserServiceTx(UserService userService, PlatformTransactionManager transactionManager) {
+        this.userService = userService;
+        this.transactionManager = transactionManager;
+    }
+
+    @Override
+    public void add(User user) {
+        userService.add(user);
+    }
+
+    @Override
+    public void upgradeLevels() {
+        TransactionStatus transactionStatus = transactionManager.getTransaction(new DefaultTransactionDefinition());
+
+        try {
+            userService.upgradeLevels();
+            transactionManager.commit(transactionStatus);
+        } catch (RuntimeException e) {
+            transactionManager.rollback(transactionStatus);
+            throw e;
+        }
+    }
+}
